Make sure validation messages are displayed, and colors are added
